% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiobe.R
\name{hall_of_fame}
\alias{hall_of_fame}
\title{Extract programming languages hall of fame from TIOBE index}
\usage{
hall_of_fame()
}
\value{
A data frame with the folowing columns: Year, Winner.
}
\description{
This function returns a dataframe containing programming languages hall of
fame from TIOBE index.
}
\examples{
hall_of_fame()
}
