% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiobe.R
\name{top_20}
\alias{top_20}
\title{Extract top 20 programming languages from TIOBE index}
\usage{
top_20()
}
\value{
A data frame with the folowing columns: Position this month,
Position the same month but previous year, Programming language, Ratings,
Change.
}
\description{
This function returns a dataframe containing top 20 programming languages
from TIOBE index.
}
\examples{
top_20()
}
