% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiobe.R
\name{top_50}
\alias{top_50}
\title{Extract top programming languages from TIOBE index with positions from 21 to 50}
\usage{
top_50()
}
\value{
A data frame with the following columns: Position, Programming
language, Ratings.
}
\description{
This function returns a dataframe containing top programming languages
from TIOBE index with positions from 21 to 50.
}
\examples{
top_50()
}
